+++ optimlol_api/endpoints/facades/championDataFacade.js
@@ -1,0 +1,27 @@
+var q = require('q');
+var _staticDataProvider = null;
+
+var _getChampionData = function(region) {
+	var deferred = q.defer();
+	_staticDataProvider.getStaticData(region, 'champions')
+		.then(function(championData) {
+			deferred.resolve(championData);
+		})
+		.then(function(error) {
+			deferred.reject(error);
+		});
+
+	return deferred.promise;
+}
+
+var _init = function() {
+	var StaticDataProvider = require('../../persistence/dataProviders/staticDataProvider');
+	_staticDataProvider = new StaticDataProvider();
+	_staticDataProvider.init();
+}
+
+module.exports = function() {
+	var self = this;
+	self.init = _init;
+	self.getChampionData = _getChampionData;
+}
